move of withAplha
